+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 <!-- CHANGELOG NEWEST -->
 
+- REFACTOR if no sort order is set on a query, use a better matching sort order and not just the primary key.
+
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->